Remove "artificial" TrimmedExceptions thrown by AbstractQueuedStreamView
In case that a transaction's address is earlier than the last
checkpoint's start address, then ignore the checkpoint and
permit the backpointer chasing to continue backward.  If an
actual trimmed address is encountered, that's OK, we will
abort the transaction.

Fix up unit tests, undoCkpointTest and CheckpointWriterInterleavedTest,
to check for correct behavior under this new regime.